+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ## _[Unreleased]_
 
+- feat: integrate "coffee" engine
 - refactor: move ggez impl into separate module
 - feat: send events to plugins
 - refactor: simplify plugin state transfer